energy-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
python-dotenv==1.0.1
redis>=5.0.0
psycopg2-binary
